+++ console/services/market_app_service.py
@@ -57,13 +57,7 @@ mnt_service = AppMntService()
 
 class MarketAppService(object):
     @staticmethod
-    def upgrade(tenant,
-                region_name,
-                user,
-                upgrade_group_id,
-                version,
-                component_keys=None,
-                is_deploy=False):
+    def upgrade(tenant, region_name, user, upgrade_group_id, version, component_keys=None, is_deploy=False):
         """
         Upgrade application market applications
         """
@@ -95,8 +89,7 @@ class MarketAppService(object):
             app_templates = json.loads(market_app_version.app_template)
             apps = app_templates["apps"]
             tenant_service_group = self.__create_tenant_service_group(
-                region_name, tenant.tenant_id, group_id, market_app.app_id, market_app_version.version,
-                market_app.app_name)
+                region_name, tenant.tenant_id, group_id, market_app.app_id, market_app_version.version, market_app.app_name)
             # install plugin for tenant
             plugins = app_templates.get("plugins", [])
             if plugins:
@@ -146,14 +139,11 @@ class MarketAppService(object):
                 items = [{"item_key": key, "item_value": config_items[key]} for key in config_items]
                 try:
                     app_config_group_service.create_config_group(group_id, config_group["name"], items,
-                                                                 config_group["injection_type"], True, component_ids,
-                                                                 region,
+                                                                 config_group["injection_type"], True, component_ids, region,
                                                                  tenant.tenant_name)
                 except ErrAppConfigGroupExists:
-                    app_config_group_service.create_config_group(group_id, config_group["name"] + "-" + make_uuid()[:4],
-                                                                 items,
-                                                                 config_group["injection_type"], True, component_ids,
-                                                                 region,
+                    app_config_group_service.create_config_group(group_id, config_group["name"] + "-" + make_uuid()[:4], items,
+                                                                 config_group["injection_type"], True, component_ids, region,
                                                                  tenant.tenant_name)
 
             # create plugin for component
@@ -236,8 +226,7 @@ class MarketAppService(object):
             new_service_list = self.__create_region_services(tenant, region_name, user, service_list)
 
             # config groups
-            config_groups = market_app_template["app_config_groups"] if market_app_template.get(
-                "app_config_groups") else []
+            config_groups = market_app_template["app_config_groups"] if market_app_template.get("app_config_groups") else []
             for config_group in config_groups:
                 component_ids = []
                 for service_key in config_group.get("component_keys", []):
@@ -264,8 +253,7 @@ class MarketAppService(object):
                     component_ids.extend(old_cgroup_service_ids)
                     items.extend(old_cgroup_items)
 
-                    app_config_group_service.update_config_group(region_name, group_id, config_group["name"], items,
-                                                                 True,
+                    app_config_group_service.update_config_group(region_name, group_id, config_group["name"], items, True,
                                                                  component_ids, tenant.tenant_name)
 
             # open plugin for component
@@ -397,8 +385,7 @@ class MarketAppService(object):
                 for item in dep_mnts:
                     dep_service = key_service_map.get(item["service_share_uuid"])
                     if not dep_service:
-                        logger.info(
-                            "Service share uuid: {}; dependent service not found".format(item["service_share_uuid"]))
+                        logger.info("Service share uuid: {}; dependent service not found".format(item["service_share_uuid"]))
                         continue
                     dep_app = app_map.get(item["service_share_uuid"])
                     if not dep_app:
@@ -438,12 +425,10 @@ class MarketAppService(object):
                         data["switch"] = True
                         data["version_id"] = build_version
                         data.update(region_config)
-                        app_plugin_service.create_service_plugin_relation(tenant.tenant_id, service.service_id,
-                                                                          plugin_id,
+                        app_plugin_service.create_service_plugin_relation(tenant.tenant_id, service.service_id, plugin_id,
                                                                           build_version)
 
-                        region_api.install_service_plugin(service.service_region, tenant.tenant_name,
-                                                          service.service_alias,
+                        region_api.install_service_plugin(service.service_region, tenant.tenant_name, service.service_alias,
                                                           data)
 
         except Exception as e:
@@ -455,8 +440,7 @@ class MarketAppService(object):
         except ServiceHandleException as e:
             logger.warning("plugin data: {0}; failed to create plugin: {1}".format(plugins, e.msg))
 
-    def __save_service_config_values(self, service, plugin_id, build_version, service_plugin_config_vars,
-                                     old_new_id_map):
+    def __save_service_config_values(self, service, plugin_id, build_version, service_plugin_config_vars, old_new_id_map):
         config_list = []
 
         for config in service_plugin_config_vars:
@@ -614,8 +598,7 @@ class MarketAppService(object):
                 for dep_key in dep_keys:
                     if dep_key["dep_service_key"] in key_service_map:
                         dep_service = key_service_map[dep_key["dep_service_key"]]
-                        code, msg, d = app_relation_service.add_service_dependency(tenant, ts, dep_service.service_id,
-                                                                                   True)
+                        code, msg, d = app_relation_service.add_service_dependency(tenant, ts, dep_service.service_id, True)
                         if code != 200:
                             logger.error("save component dependency relation error {0}".format(msg))
 
@@ -830,22 +813,21 @@ class MarketAppService(object):
             extend_info["market_name"] = market_name
         service_source_params = {
             "team_id":
-                ts.tenant_id,
+            ts.tenant_id,
             "service_id":
-                ts.service_id,
+            ts.service_id,
             "user_name":
-                "",
+            "",
             "password":
-                "",
+            "",
             "extend_info":
-                json.dumps(extend_info),
+            json.dumps(extend_info),
             "group_key":
-                app_id,
+            app_id,
             "version":
-                version,
+            version,
             "service_share_uuid":
-                component.get("service_share_uuid") if component.get("service_share_uuid", None) else component.get(
-                    "service_key")
+            component.get("service_share_uuid") if component.get("service_share_uuid", None) else component.get("service_key")
         }
         service_source_repo.create_service_source(**service_source_params)
 
@@ -873,8 +855,7 @@ class MarketAppService(object):
             count = rainbond_app_repo.get_rainbond_app_total_count(eid, scope, teams, app_name, tag_names, need_install)
         else:
             # default scope is enterprise
-            apps = rainbond_app_repo.get_rainbond_app_in_enterprise_by_query(eid, scope, app_name, tag_names, page,
-                                                                             page_size,
+            apps = rainbond_app_repo.get_rainbond_app_in_enterprise_by_query(eid, scope, app_name, tag_names, page, page_size,
                                                                              need_install)
             count = rainbond_app_repo.get_rainbond_app_total_count(eid, scope, None, app_name, tag_names, need_install)
         if not apps:
@@ -924,8 +905,7 @@ class MarketAppService(object):
     # patch rainbond app versions
     def _patch_rainbond_app_versions(self, eid, apps, is_complete):
         app_ids = [app.app_id for app in apps]
-        versions = rainbond_app_repo.get_rainbond_app_version_by_app_ids(eid, app_ids, is_complete,
-                                                                         rm_template_field=True)
+        versions = rainbond_app_repo.get_rainbond_app_version_by_app_ids(eid, app_ids, is_complete, rm_template_field=True)
         if not versions:
             return
 
@@ -948,7 +928,7 @@ class MarketAppService(object):
             # If the versions are the same, take the last version information
             app_with_versions[version.app_id][version_info["version"]] = version_info
             if version_info["version"] in app_release_ver_nums[
-                version.app_id] or version_info["version"] in app_not_release_ver_nums[version.app_id]:
+                    version.app_id] or version_info["version"] in app_not_release_ver_nums[version.app_id]:
                 continue
             if version_info["dev_status"] == "release":
                 app_release_ver_nums[version.app_id].append(version_info["version"])
@@ -1030,8 +1010,7 @@ class MarketAppService(object):
         tenants = team_repo.get_teams_by_enterprise_id(enterprise_id)
         tenant_names = [t.tenant_name for t in tenants]
         # 获取企业分享的应用，并且排除返回在团队内的
-        return rainbond_app_repo.get_current_enter_visable_apps(enterprise_id).filter(
-            share_team__in=tenant_names).exclude(
+        return rainbond_app_repo.get_current_enter_visable_apps(enterprise_id).filter(share_team__in=tenant_names).exclude(
             scope="team")
 
     def get_public_market_shared_apps(self, enterprise_id):
@@ -1072,8 +1051,7 @@ class MarketAppService(object):
             try:
                 market = app_market_service.get_app_market_by_name(
                     tenant.enterprise_id, extend_info.get("market_name"), raise_exception=True)
-                resp = app_market_service.get_market_app_model_version(market, service_source.group_key,
-                                                                       service_source.version)
+                resp = app_market_service.get_market_app_model_version(market, service_source.group_key, service_source.version)
                 if not resp:
                     raise app_not_found
             except region_api.CallApiError as e:
@@ -1140,11 +1118,9 @@ class MarketAppService(object):
                 scope="goodrain", source="market", group_name__icontains=app_name)
         if is_complete:
             if is_complete == "true":
-                return rainbond_app_repo.get_all_rainbond_apps().filter(scope="goodrain", source="market",
-                                                                        is_complete=True)
+                return rainbond_app_repo.get_all_rainbond_apps().filter(scope="goodrain", source="market", is_complete=True)
             else:
-                return rainbond_app_repo.get_all_rainbond_apps().filter(scope="goodrain", source="market",
-                                                                        is_complete=False)
+                return rainbond_app_repo.get_all_rainbond_apps().filter(scope="goodrain", source="market", is_complete=False)
         return rainbond_app_repo.get_all_rainbond_apps().filter(scope="goodrain", source="market")
 
     def list_upgradeable_versions(self, tenant, service):
@@ -1155,10 +1131,8 @@ class MarketAppService(object):
             install_from_cloud = component_source.is_install_from_cloud()
             if install_from_cloud and market_name:
                 market = app_market_repo.get_app_market_by_name(tenant.enterprise_id, market_name, raise_exception=True)
-            return self.__get_upgradeable_versions(tenant.enterprise_id, component_source.group_key,
-                                                   component_source.version,
-                                                   component_source.get_template_update_time(), install_from_cloud,
-                                                   market)
+            return self.__get_upgradeable_versions(tenant.enterprise_id, component_source.group_key, component_source.version,
+                                                   component_source.get_template_update_time(), install_from_cloud, market)
         return []
 
     def get_enterprise_access_token(self, enterprise_id, access_target):
@@ -1206,8 +1180,7 @@ class MarketAppService(object):
                 if market:
                     app_model, _ = app_market_service.cloud_app_model_to_db_model(market, app_model_key, version=None)
             else:
-                app_model, _ = rainbond_app_repo.get_rainbond_app_and_version(tenant.enterprise_id, app_model_key,
-                                                                              version)
+                app_model, _ = rainbond_app_repo.get_rainbond_app_and_version(tenant.enterprise_id, app_model_key, version)
             if not app_model:
                 continue
             dat = {
@@ -1229,8 +1202,7 @@ class MarketAppService(object):
             }
             not_upgrade_record = upgrade_service.get_app_not_upgrade_record(tenant.tenant_id, app_id, app_model_key)
             versions = self.__get_upgradeable_versions(tenant.enterprise_id, app_model_key, version,
-                                                       component_source.get_template_update_time(), install_from_cloud,
-                                                       market)
+                                                       component_source.get_template_update_time(), install_from_cloud, market)
             dat.update({
                 'current_version': version,
                 'can_upgrade': bool(versions),
@@ -1298,8 +1270,7 @@ class MarketAppService(object):
     def get_models_upgradeable_version(self, enterprise_id, app_model_key, app_id, upgrade_group_id):
         current_version, current_version_update_time, install_from_cloud, market = self.get_current_version(
             enterprise_id, app_model_key, app_id, upgrade_group_id)
-        return self.__get_upgradeable_versions(enterprise_id, app_model_key, current_version,
-                                               current_version_update_time,
+        return self.__get_upgradeable_versions(enterprise_id, app_model_key, current_version, current_version_update_time,
                                                install_from_cloud, market)
 
     def delete_rainbond_app_all_info_by_id(self, enterprise_id, app_id):
